requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
tensorflow~=2.4.1
numpy==1.19.2
Pillow~=8.2.0
fastapi~=0.63.0
pydantic~=1.8.1
python-multipart